refactor(commands): extract /teleport command into teleport.rs
